remove false positive overlap error on simple annotations without assign
closes #987
