+++ tests/test_visitors/test_ast/test_blocks/test_overlap/test_import_block.py
@@ -52,7 +52,7 @@ def test_import_block_overlap(
     assert_error_text,
     parse_ast_tree,
     import_statement,
-    assign_statement,
+    assign_and_annotation_statement,
     context,
     variable_name,
     default_options,
@@ -61,7 +61,7 @@ def test_import_block_overlap(
     """Ensures that overlaping variables exist."""
     code = context.format(
         import_statement.format(variable_name),
-        assign_statement.format(variable_name),
+        assign_and_annotation_statement.format(variable_name),
     )
     tree = parse_ast_tree(mode(code))
 
@@ -127,7 +127,7 @@ def test_import_block_correct(
     assert_errors,
     parse_ast_tree,
     import_statement,
-    assign_statement,
+    assign_and_annotation_statement,
     context,
     first_name,
     second_name,
@@ -137,7 +137,7 @@ def test_import_block_correct(
     """Ensures that different variables do not overlap."""
     code = context.format(
         import_statement.format(first_name),
-        assign_statement.format(second_name),
+        assign_and_annotation_statement.format(second_name),
     )
     tree = parse_ast_tree(mode(code))
