+++ tests/test_visitors/test_ast/test_blocks/test_overlap/test_loop_blocks.py
@@ -63,7 +63,7 @@ def test_for_block_overlap(
     assert_error_text,
     parse_ast_tree,
     for_statement,
-    assign_statement,
+    assign_and_annotation_statement,
     context,
     variable_name,
     default_options,
@@ -72,7 +72,7 @@ def test_for_block_overlap(
     """Ensures that overlaping variables exist."""
     code = context.format(
         for_statement.format(variable_name),
-        assign_statement.format(variable_name),
+        assign_and_annotation_statement.format(variable_name),
     )
     tree = parse_ast_tree(mode(code))
 
@@ -138,7 +138,7 @@ def test_for_block_correct(
     assert_errors,
     parse_ast_tree,
     for_statement,
-    assign_statement,
+    assign_and_annotation_statement,
     context,
     first_name,
     second_name,
@@ -148,7 +148,7 @@ def test_for_block_correct(
     """Ensures that different variables do not overlap."""
     code = context.format(
         for_statement.format(first_name),
-        assign_statement.format(second_name),
+        assign_and_annotation_statement.format(second_name),
     )
     tree = parse_ast_tree(mode(code))
