+++ tests/test_visitors/test_ast/test_blocks/test_overlap/test_with_block.py
@@ -67,7 +67,7 @@ def test_with_block_overlap(
     assert_error_text,
     parse_ast_tree,
     with_statement,
-    assign_statement,
+    assign_and_annotation_statement,
     context,
     variable_name,
     default_options,
@@ -76,7 +76,7 @@ def test_with_block_overlap(
     """Ensures that overlaping variables exist."""
     code = context.format(
         with_statement.format(variable_name),
-        assign_statement.format(variable_name),
+        assign_and_annotation_statement.format(variable_name),
     )
     tree = parse_ast_tree(mode(code))
 
@@ -146,7 +146,7 @@ def test_with_block_correct(
     assert_errors,
     parse_ast_tree,
     with_statement,
-    assign_statement,
+    assign_and_annotation_statement,
     context,
     first_name,
     second_name,
@@ -156,7 +156,7 @@ def test_with_block_correct(
     """Ensures that different variables do not overlap."""
     code = context.format(
         with_statement.format(first_name),
-        assign_statement.format(second_name),
+        assign_and_annotation_statement.format(second_name),
     )
     tree = parse_ast_tree(mode(code))
